Add nbreParts component
